Refactor game cards with new GameCardContainer component
The old method of assigning fixed dimensions to the game card has been refactored for better reusability. A new `GameCardContainer` component was introduced to manage card dimensions and styling that were previously handled by individual game cards. This also enhances the appearance of the skeleton loading state in the Game Grid component.