Add GenreList component to aside area in App
Added a new GenreList component to the 'aside' area in App component. A new hook called 'useGenres' has been implemented to fetch genres from the server, and then it is used inside the GenreList component to display a list of genres. It will improve the Website's side navigation by providing a list of available game genres.