Add Emoji component and rating feature
An Emoji component has been added with different images associated with different rating numbers. In addition, rating properties have been incorporated into the GameCard and useGames components. The Emoji component uses a rating prop to display an appropriate emoji for each game depending on the top rating of the game.