Refactor code to use Zustand state management
Zustand state management library was added to manage global state. A new store was created to store the game query parameters. The components were refactored to use this store instead of passing around props. This simplifies the component tree and makes state changes more predictable. A new dependency, Zustand, is also added to the 'package.json' and 'package-lock.json' files.