Implement react-router and refactor application structure
This update incorporates react-router-dom into the application for routing functionality. The files `App.css` and `App.tsx` are deleted and their logic is moved to new respective components in `HomePage.tsx`. Furthermore, `Layout.tsx` has been added to handle the overall layout including the navigation bar. Additionally, 'GameDetailPage.tsx' is introduced for game details, and routes are declared in 'routes.tsx'. Lastly, 'react-router-dom' is added to the dependencies in 'package.json' and 'package-lock.json'.