Refactor code to move entity interfaces to separate files
Implemented a code refactor where the `Game`, `Genre`, and `Platform` interfaces have been moved out of the hooks and into separate files in the `entities` directory. This streamlines the code, making it cleaner and easier to navigate, as well as centralizing the definitions of these objects. The hooks and components that used these interfaces have been updated to import them from their new location.